pyproject.toml
[tool.poetry]
name = "django-advanced-logging"
version = "1.0.0"
description = "Sistema de logging profesional y escalable para proyectos Django con soporte para PostgreSQL - Uso Privado Empresa"
authors = ["Equipo Python <python-team@tu-empresa.com>"]
license = "Proprietary"  # O "Private" - NO es código abierto
readme = "README_PRIVATE.md"
# URLs internas de la empresa
homepage = "https://git.tu-empresa.com/python/django-advanced-logging"
repository = "https://git.tu-empresa.com/python/django-advanced-logging"
documentation = "https://wiki.tu-empresa.com/python/django-advanced-logging"
packages = [
    { include = "django_advanced_logging" }
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
# Django es OBLIGATORIO
Django = ">=3.2"
# PostgreSQL obligatorio
psycopg2-binary = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
mypy = "^1.5.0"
django-stubs = "^4.2.0"
pre-commit = "^3.3.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.0"
mkdocs-material = "^9.2.0"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}

# Django y PostgreSQL son obligatorios, no hay extras opcionales

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip_glob = ["*/migrations/*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "-ra -q --cov=django_advanced_logging --cov-report=term-missing"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "tests.settings"

[tool.coverage.run]
source = ["django_advanced_logging"]
omit = ["*/migrations/*", "*/tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false